Self-Update Framework and File Deployment
The system checks for updates, displays a banner if updates exist,
provides a link to run the update, copies the files (success with views
and dll files so far), updates the version in the web.config, and
provides a message saying update succeeded. Currently logging the update
activities in custom tables, and all error related messages using
log4net.

Next steps is to customize the response, test running multiple updates
in series, reverting to previous versions.

Once the framework is solid I will need to implement the SQL script
update, web.config update (other than version), and a delete command to
remove files.

Once the framework is solid I need to move it into it's own class
library and potentially post on NuGet.